requirements.txt
appdirs==1.4.0
click==6.7
Flask==0.12
Flask-Login==0.4.0
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
packaging==16.8
passlib==1.7.1
pyparsing==2.1.10
python-magic==0.4.12
PyYAML==3.12
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.5
Werkzeug==0.11.15
